chore(solhint): config solidity linting

+++ src/ethereum/compound/contracts/ICEther.sol
@@ -1,17 +1,35 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.6.0 <= 0.9.0;
+pragma solidity >=0.6.0 <=0.9.0;
 
 interface ICEther {
     function mint() external payable;
-    function borrow(uint borrowAmount) external returns (uint);
-    function redeem(uint redeemTokens) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+
     function repayBorrow() external payable;
+
     function repayBorrowBehalf(address borrower) external payable;
-    function borrowBalanceCurrent(address account) external returns (uint);
+
+    function borrowBalanceCurrent(address account) external returns (uint256);
+
     function borrowBalanceStored(address account) external view returns (uint256);
-    function balanceOfUnderlying(address account) external returns (uint);
+
+    function balanceOfUnderlying(address account) external returns (uint256);
+
     function balanceOf(address owner) external view returns (uint256);
-    function getAccountSnapshot(address account) external view returns (uint, uint, uint, uint);
+
+    function getAccountSnapshot(address account)
+        external
+        view
+        returns (
+            uint256,
+            uint256,
+            uint256,
+            uint256
+        );
 }
